Fixed state reset on forget password and validate email screens.
