+++ src/RegistrationAndLogin/RegistrationPage.jsx
@@ -1,102 +1,34 @@
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import SubmissionButton from "../components/SubmissionButton";
+import RegistrationDisplay from "./components/RegistrationDisplay";
+import ValidateEmailDisplay from "./components/ValidateEmailDisplay";
 
 function RegistrationPage() {
   const [email, setEmail] = useState("");
   const [password, setPassword] = useState("");
   const [password2, setPassword2] = useState("");
+  const [code, setCode] = useState("");
+  const [page, setPage] = useState("registration");
 
   const navigate = useNavigate();
 
-  function changePassword(event) {
-    setPassword(event.target.value);
+  let registrationPage = false;
+  if (page === "registration") {
+    registrationPage = true;
   }
 
-  function changePassword2(event) {
-    setPassword2(event.target.value);
-  }
-
-  function changeEmail(event) {
-    setEmail(event.target.value);
-  }
-
-  let canRegister = false;
-
-  if (email.length > 0 && password.length > 0 && password2.length > 0) {
-    canRegister = true;
-  }
-
-  return (
-    <div className="min-h-screen bg-gray-100 flex flex-col items-center justify-center font-[Nunito] px-4">
-      <div className="bg-white p-8 rounded-lg shadow-md w-full max-w-md">
-        <h1 className="text-4xl font-bold text-center mb-6">Register</h1>
-
-        <label
-          className="block text-sm font-medium text-gray-700 mb-1"
-          htmlFor="email"
-        >
-          Email
-        </label>
-        <input
-          id="email"
-          type="text"
-          autoComplete="email"
-          placeholder="Type your email"
-          className="w-full px-4 py-2 mb-4 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-          value={email}
-          onChange={(event) => changeEmail(event)}
-        />
-
-        <label
-          className="block text-sm font-medium text-gray-700 mb-1"
-          htmlFor="password"
-        >
-          Password
-        </label>
-        <input
-          id="password"
-          type="password"
-          autoComplete="off"
-          placeholder="Type your password"
-          className="w-full px-4 py-2 mb-6 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-          value={password}
-          onChange={(event) => changePassword(event)}
-        />
-
-        <label
-          className="block text-sm font-medium text-gray-700 mb-1"
-          htmlFor="password"
-        >
-          Confirm Password
-        </label>
-        <input
-          id="password2"
-          type="password"
-          autoComplete="off"
-          placeholder="Retype your password"
-          className="w-full px-4 py-2 mb-6 border border-gray-300 rounded-md focus:outline-none focus:ring-2 focus:ring-blue-500"
-          value={password2}
-          onChange={(event) => changePassword2(event)}
-        />
-
-        <SubmissionButton
-          isActive={canRegister}
-          onClick={() => navigate("/validate-email")}
-        >
-          Register
-        </SubmissionButton>
-
-        <div className="flex justify-center mt-4">
-          <button
-            className="text-sm text-blue-600 hover:underline"
-            onClick={() => navigate("/login")}
-          >
-            Already a member? Log in here
-          </button>
-        </div>
-      </div>
-    </div>
+  return registrationPage ? (
+    <RegistrationDisplay
+      email={email}
+      setEmail={setEmail}
+      password={password}
+      setPassword={setPassword}
+      password2={password2}
+      setPassword2={setPassword2}
+      setPage={setPage}
+    />
+  ) : (
+    <ValidateEmailDisplay code={code} setCode={setCode} setPage={setPage} />
   );
 }
